containers/lib/requirements.txt
opentelemetry-api
opentelemetry-propagator-aws-xray
